refactor RTCore, make Filters class
