+++ ringtail/outputmanager.py
@@ -1,0 +1,307 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Ringtail output manager
+#
+
+from .exceptions import OutputError
+import logging
+import typing
+import json
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib import cm
+from matplotlib import colors
+from rdkit.Chem import SDWriter
+from meeko import RDKitMolCreate
+
+
+class OutputManager:
+    """Class for creating outputs
+
+    Attributes:
+        log (string): name for log file
+        export_sdf_path (str): path for exporting SDF molecule files
+
+    """
+
+    def __init__(self, log_file="output_log.txt", export_sdf_path="", _stop_at_defaults=False):
+        """Initialize OutputManager object and create log file
+
+        Args:
+            log_file (string): name for log file
+            export_sdf_path (string): path for exporting sdf files
+        """
+
+        self.log = log_file
+        self.export_sdf_path = export_sdf_path
+        if _stop_at_defaults:
+            return
+    
+    @classmethod
+    def get_defaults(cls):
+        return cls(_stop_at_defaults=True).__dict__
+
+    @classmethod
+    def get_default_types(cls):
+        return typing.get_type_hints(cls.__init__)
+
+    def plot_all_data(self, binned_data):
+        """takes dictionary of binned data where key is the
+        coordinates of the bin and value is the number of points in that bin.
+        Adds to scatter plot colored by value
+
+        Args:
+            binned_data (dict): Keys are tuples of key and y value for bin.
+                Value is the count of points falling into that bin.
+        """
+        try:
+            # gather data
+            energies = []
+            leffs = []
+            bin_counts = []
+            for data_bin in binned_data.keys():
+                energies.append(data_bin[0])
+                leffs.append(data_bin[1])
+                bin_counts.append(binned_data[data_bin])
+
+            # start with a square Figure
+            fig = plt.figure()
+
+            gs = fig.add_gridspec(
+                2,
+                2,
+                width_ratios=(7, 2),
+                height_ratios=(2, 7),
+                left=0.1,
+                right=0.9,
+                bottom=0.1,
+                top=0.9,
+                wspace=0.05,
+                hspace=0.05,
+            )
+
+            self.ax = fig.add_subplot(gs[1, 0])
+            ax_histx = fig.add_subplot(gs[0, 0], sharex=self.ax)
+            ax_histy = fig.add_subplot(gs[1, 1], sharey=self.ax)
+            fig.colorbar(
+                mappable=cm.ScalarMappable(
+                    colors.Normalize(vmin=min(bin_counts), vmax=max(bin_counts))
+                ),
+                label="Scatterplot bin count",
+            )
+            self.ax.set_xlabel("Best docking score / kcal/mol")
+            self.ax.set_ylabel("Best Ligand Efficiency")
+        except Exception as e:
+            raise OutputError("Error occurred while initializing plot") from e
+
+        self.scatter_hist(energies, leffs, bin_counts, self.ax, ax_histx, ax_histy)
+
+    def plot_single_point(self, x, y, color="black"):
+        """Add point to scatter plot with given x and y coordinates and color.
+
+        Args:
+            x (float): x coordinate
+            y (float): y coordinate
+            color (str, optional): Color for point. Default black.
+        """
+        try:
+            self.ax.scatter([x], [y], c=color)
+        except Exception as e:
+            raise OutputError("Error occurred while plotting") from e
+
+    def save_scatterplot(self):
+        """
+        Saves current figure as scatter.png
+        """
+        try:
+            plt.savefig("scatter.png", bbox_inches="tight")
+            plt.close()
+        except Exception as e:
+            raise OutputError("Error while saving figure") from e
+
+    def write_log(self, lines):
+        """Writes lines from results iterable into log file
+
+        Args:
+            lines (iterable): Iterable with tuples of data for
+                writing into log
+        """
+        try:
+            for line in lines:
+                logging.info(line)
+                self._write_log_line(
+                    str(line).replace("(", "").replace(")", "")
+                )  # strip parens from line, which is natively a tuple
+            self._write_log_line("***************\n")
+        except Exception as e:
+            raise OutputError("Error occurred during log writing") from e
+
+    def _write_log_line(self, line):
+        """write a single row to the log file
+
+        Args:
+            line (string): Line to write to log
+        """
+        try:
+            with open(self.log, "a") as f:
+                f.write(line)
+                f.write("\n")
+        except Exception as e:
+            raise OutputError(f"Error writing line {line} to log") from e
+
+    def log_num_passing_ligands(self, number_passing_ligands):
+        """
+        Write the number of ligands which pass given filter to log file
+
+        Args:
+            number_passing_ligands (int): number of ligands that passed filter
+        """
+        try:
+            with open(self.log, "a") as f:
+                f.write("\n")
+                f.write(
+                    "Number passing ligands: {num} \n".format(
+                        num=str(number_passing_ligands)
+                    )
+                )
+                f.write("---------------\n")
+        except Exception as e:
+            raise OutputError("Error writing number of passing ligands in log") from e
+
+    def write_results_bookmark_to_log(self, bookmark_name):
+        """Write the name of the result bookmark into log
+
+        Args:
+            bookmark_name (string): name of current results' bookmark in db
+        """
+        try:
+            with open(self.log, "a") as f:
+                f.write("\n")
+                f.write(f"Result bookmark name: {bookmark_name}\n")
+        except Exception as e:
+            raise OutputError("Error writing bookmark name to log") from e
+
+    def write_out_mol(self, ligname, mol, flexres_mols, properties):
+        """writes out given mol as sdf
+
+        Args:
+            ligname (string): name of ligand that will be used to
+                name output SDF file
+            mol (RDKit mol object): RDKit molobject to be written to SDF
+            flexres_mols (list): dictionary of rdkit molecules for flexible residues
+            properties (dict): dictionary of list of properties to add to mol before writing
+        """
+        try:
+            filename = self.export_sdf_path + ligname + ".sdf"
+            mol_flexres_list = [mol]
+            mol_flexres_list += flexres_mols
+            mol = RDKitMolCreate.combine_rdkit_mols(mol_flexres_list)
+            # convert properties to strings as needed
+            for k, v in properties.items():
+                if isinstance(v, list):
+                    v = json.dumps(v)
+                elif not isinstance(v, str):
+                    v = str(v)
+                mol.SetProp(k, v)
+
+            with SDWriter(filename) as w:
+                for conf in mol.GetConformers():
+                    w.write(mol, conf.GetId())
+
+        except Exception as e:
+            raise OutputError("Error occurred while writing SDF from RDKit Mol") from e
+
+    def create_log_file(self):
+        """
+        Initializes log file
+        """
+        try:
+            with open(self.log, "w") as f:
+                f.write("Filtered poses:\n")
+                f.write("***************\n")
+        except Exception as e:
+            raise OutputError("Error while creating log file") from e
+
+    def scatter_hist(self, x, y, z, ax, ax_histx, ax_histy):
+        """
+        Makes scatterplot with a histogram on each axis
+
+        Args:
+            x (list): x coordinates for data
+            y (list): y coordinates for data
+            z (list): z coordinates for data
+            ax (matplotlib axis): scatterplot axis
+            ax_histx (matplotlib axis): x histogram axis
+            ax_histy (matplotlib axis): y histogram axis
+        """
+        try:
+            # no labels
+            ax_histx.tick_params(axis="x", labelbottom=False)
+            ax_histy.tick_params(axis="y", labelleft=False)
+
+            # the scatter plot:
+            ax.scatter(x, y, c=z, cmap="viridis")
+
+            # now determine nice limits by hand:
+            xbinwidth = 0.25
+            ybinwidth = 0.01
+            xminlim = (int(min(x) / xbinwidth) + 3) * xbinwidth
+            xmaxlim = (int(max(x) / xbinwidth) + 3) * xbinwidth
+            yminlim = (int(min(y) / ybinwidth) + 3) * ybinwidth
+            ymaxlim = (int(max(y) / ybinwidth) + 3) * ybinwidth
+
+            xbins = np.arange(xminlim, xmaxlim + xbinwidth, xbinwidth)
+            ybins = np.arange(yminlim, ymaxlim + ybinwidth, ybinwidth)
+
+            ax_histx.hist(x, bins=xbins)
+            ax_histy.hist(y, bins=ybins, orientation="horizontal")
+        except Exception as e:
+            raise OutputError("Error occurred while adding all data to plot") from e
+
+    def write_filters_to_log(self, filters_dict, included_interactions):
+        """Takes dictionary of filters, formats as string and writes to log file
+
+        Args:
+            filters_dict (dict): dictionary of filtering options
+        """
+        try:
+            buff = ["##### PROPERTIES"]
+            for k, v in filters_dict["properties"].items():
+                if v is not None:
+                    v = "%2.3f" % v
+                else:
+                    v = " [ none ]"
+                buff.append("#  % 7s : %s" % (k, v))
+            if filters_dict["filter_ligands_flag"]:
+                buff.append("#### LIGAND FILTERS")
+                for k, v in filters_dict["ligand_filters"].items():
+                    if v is not None:
+                        if isinstance(v, list):
+                            v = ", ".join([f for f in v if f != ""])
+                    else:
+                        v = " [ none ]"
+                    buff.append("#  % 7s : %s" % (k, v))
+            buff.append("#### INTERACTIONS")
+            labels = ["~", ""]
+            for _type, info in filters_dict["interactions"].items():
+                kept_interactions = []
+                if len(info) == 0:
+                    buff.append("#  % 7s :  [ none ]" % (_type))
+                    continue
+                for interact in info:
+                    if _type + "-" + interact[0] not in included_interactions:
+                        continue
+                    else:
+                        kept_interactions.append(interact)
+                res_str = ", ".join(
+                    ["(%s)%s" % (labels[int(x[1])], x[0]) for x in kept_interactions]
+                )
+                l_str = "#  % 7s : %s" % (_type, res_str)
+                buff.append(l_str)
+
+            for line in buff:
+                self._write_log_line(line)
+
+        except Exception as e:
+            raise OutputError("Error occurred while writing filters to log") from e
