+++ ringtail/ringtailcore.py
@@ -4,24 +4,19 @@
 # Ringtail virtual screening manager
 #
 
-from cProfile import label
 import matplotlib.pyplot as plt
-from matplotlib import cm
-from matplotlib import colors
-import numpy as np
 import json
 from meeko import RDKitMolCreate
 from .storagemanager import StorageManager, StorageManagerSQLite
 from .resultsmanager import ResultsManager
 from .receptormanager import ReceptorManager
-from .exceptions import StorageError
+from .outputmanager import OutputManager
+from .filters import Filters
 from .exceptions import RTCoreError, OutputError
 from rdkit import Chem
-from rdkit.Chem import SDWriter
 import itertools
 import logging
 import os
-from typing import Union
 
 
 class RingtailCore:
@@ -46,7 +41,7 @@ class RingtailCore:
             files for insertion into database
     """
 
-    def __init__(self, **opts):
+    def __init__(self, storage_type = "sqlite", opts_dict: dict = None):
         """Initialize RingtailCore object. Will create storageManager object to serve
         as interface with database (currently implemented in SQLite).
         Will create ResultsManager to process result files.
@@ -60,32 +55,32 @@ class RingtailCore:
             out_opts (dictionary): Specified output options including data
                 fields to output, export_sdf_path, log file name
         """
-        defaults = self.get_defaults()
+        # check if given opt_dict and set attrbutes if so
+        if opts_dict is None:
+            self.storage_opts = {}
+            self.rman_opts = {}
+            self.filters_dict = {}
+            self.out_opts = {}
+        else:
+            self.storage_opts = opts_dict["storage_opts"]["values"]
+            self.rman_opts = opts_dict["rman_opts"]["values"]
+            self.filters_dict = opts_dict["filters"]["values"]
+            self.out_opts = opts_dict["out_opts"]["values"]
 
-        # unpack given opts dictionaries
-        for k, v in opts.items():
-            if k in defaults.keys():
-                defaults[k] = v
-
-        self.storage_opts = defaults["storage_opts"]["values"]
-        self.rman_opts = defaults["rman_opts"]["values"]
-        self.filters = defaults["filters"]["values"]
-        self.out_opts = defaults["out_opts"]["values"]
-
-        storage_type = self.storage_opts.pop("storage_type")
-        # confirm storage_type is not None
-        if storage_type is None:
-            raise RTCoreError(
-                "storage_type not defined and must be defined at runtime."
-            )
-
+        # set storage type
+        self.storage_type = storage_type
         storage_types = {
             "sqlite": StorageManagerSQLite,
         }
+        # confirm given storage type is implemented
+        if self.storage_type not in storage_types:
+            raise NotImplementedError(f"Given storage type {self.storage_type} is not implemented.")
+        
+        # initialize "worker" classes
         self.storageman = storage_types[storage_type](**self.storage_opts)
-        self.rman_opts["storageman"] = self.storageman
-        self.rman_opts["storageman_class"] = storage_types[storage_type]
-        self.output_manager = None
+        self.results_man = ResultsManager(storageman=self.storageman, storageman_class=storage_types[storage_type], **self.rman_opts)
+        self.output_manager = OutputManager(**self.out_opts)
+        self.filters = Filters(**self.filters_dict)
 
     @classmethod
     def get_defaults(cls, storage_type="sqlite", terse=False) -> dict:
@@ -101,283 +96,24 @@ class RingtailCore:
                 "ignore": [],
                 "types": ResultsManager.get_default_types(),
             },
-        }
-
-        # add storage type default
-        cls.__default_options["storage_opts"]["values"]["storage_type"] = storage_type
-        cls.__default_options["storage_opts"]["types"]["storage_type"] = str
-
-        out_opts = {
-            "log": "output_log.txt",
-            "export_sdf_path": None,
-            "plot": None,
-            "outfields": "e",
-            "no_print": True,
-            "export_bookmark_csv": None,
-            "export_query_csv": None,
-            "export_bookmark_db": None,
-            "data_from_bookmark": None,
-            "filter_bookmark": None,
-        }
-
-        out_types = {
-            "log": str,
-            "export_sdf_path": str,
-            "plot": bool,
-            "outfields": str,
-            "no_print": bool,
-            "export_bookmark_csv": str,
-            "export_query_csv": str,
-            "export_bookmark_db": str,
-            "data_from_bookmark": str,
-            "filter_bookmark": str,
-        }
-
-        filters = {
-            "properties": {
-                "eworst": None,
-                "ebest": None,
-                "leworst": None,
-                "lebest": None,
-                "score_percentile": None,
-                "le_percentile": None,
+            "out_opts": {
+                "values": OutputManager.get_defaults(),
+                "ignore": [],
+                "types": OutputManager.get_default_types(),
             },
-            "interactions": {"V": [], "H": [], "R": []},
-            "interactions_count": [],
-            "ligand_filters": {"N": [], "S": [], "F": "OR", "X": []},
-            "filter_ligands_flag": False,
-            "max_miss": 0,
-            "react_any": None,
-        }
-
-        filter_types = {
-            "properties": {
-                "eworst": float,
-                "ebest": float,
-                "leworst": float,
-                "lebest": float,
-                "score_percentile": float,
-                "le_percentile": float,
+            "filters_dict": {
+                "values": Filters.get_defaults(),
+                "ignore": [],
+                "types": Filters.get_default_types(),
             },
-            "interactions": dict,
-            "interactions_count": list,
-            "ligand_filters": dict,
-            "filter_ligands_flag": bool,
-            "max_miss": int,
-            "react_any": bool,
         }
 
         defaults = cls.__default_options.copy()
         if terse:
             for group, opt in defaults.items():
                 defaults[group] = {k: v for k, v in opt.items() if not k == "ignore"}
-        defaults["out_opts"] = {"values": out_opts, "ignore": [], "types": out_types}
-        defaults["filters"] = {"values": filters, "ignore": [], "types": filter_types}
 
         return defaults
-
-    @classmethod
-    def set_opts(cls, opts_dict: dict, option_names: Union[str, list], values) -> dict:
-        """Given dictionary of options, set given option_name(s) to given value(s). Indices of option_names and values should correspond. Will not set filters
-
-        Args:
-            opts_dict (dict): dictionary of RTCore options. Follow format of dict returned by cls.get_defaults()
-            option_names (string or list): name(s) of option to change
-            values: values to set option_name(s) to
-        """
-        # format inputs into lists if not already
-        if type(option_names) != list:
-            if type(option_names) == str:
-                option_names = [option_names]
-                values = [values]
-            else:
-                raise TypeError(f"Expecting string or list for option_names. Given {type(option_names)}")
-        # set file_source options
-        for idx, option_name in enumerate(option_names):
-            if option_name == "file" or option_name == "file_list":
-                opts_dict["rman_opts"]["values"]["file_sources"][option_name] = values[idx]
-            elif option_name == "path" or option_name == "pattern" or option_name == "recursive":
-                opts_dict["rman_opts"]["values"]["file_sources"]["file_path"][option_name] = values[idx]
-            # set anything else
-            else:
-                found = False
-                for opts_subdict in opts_dict:
-                    if opts_subdict == "filters":
-                        continue
-                    if option_name in opts_dict[opts_subdict]["values"]:
-                        opts_dict[opts_subdict]["values"][option_name] = values[idx]
-                        found = True
-
-                if not found:
-                    raise RTCoreError(f"Given option {option_name} not found. Please check spelling try setting again.")
-
-        return opts_dict
-
-    @classmethod
-    def get_opts(cls, opts_dict: dict, option_names: Union[str, list]) -> list:
-        """Given dictionary of options and option_names, get value. Will not get filters
-
-        Args:
-            opts_dict (dict): dictionary of RTCore options. Should follow format of dict returned by cls.get_defaults()
-            option_names (str): name of option to change
-
-        Returns:
-            list: list of option values for given option types
-        """
-        if type(option_names) != list:
-            if type(option_names) == str:
-                option_names = [option_names]
-            else:
-                raise TypeError(f"Expecting string or list for option_names. Given {type(option_names)}")
-        opts = []
-        for option_name in option_names:
-            if option_name == "file" or option_name == "file_list":
-                opts.append(opts_dict["rman_opts"]["values"]["file_sources"][option_name])
-            elif option_name == "path" or option_name == "pattern" or option_name == "recursive":
-                opts.append(opts_dict["rman_opts"]["values"]["file_sources"]["file_path"][option_name])
-            else:
-                for opts_subdict in opts_dict:
-                    if opts_subdict == "filters":
-                        continue
-                    if option_name in opts_dict[opts_subdict]["values"]:
-                        opts.append(opts_dict[opts_subdict]["values"][option_name])
-                # will only enter if we failed to find and return the option value
-                logging.warning(f"Given option {option_name} not found.")
-
-        return opts
-
-    @classmethod
-    def get_opt_types(cls, opts_dict: dict, option_names: Union[str, list]) -> list:
-        """Given dictionary of options and option_names, get type of option_names. Will not get filters
-
-        Args:
-            opts_dict (dict): dictionary of RTCore options. Should follow format of dict returned by cls.get_defaults()
-            option_names (str or list): name(s) of option to get type for
-
-        Returns:
-            list: list of option types for given option types
-        """
-        if type(option_names) != list:
-            if type(option_names) == str:
-                option_names = [option_names]
-            else:
-                raise TypeError(f"Expecting string or list for option_names. Given {type(option_names)}")
-        types = []
-        for option_name in option_names:
-            if option_name == "file" or option_name == "file_list":
-                types.append(opts_dict["rman_opts"]["types"]["file_sources"][option_name])
-            elif option_name == "path" or option_name == "pattern" or option_name == "recursive":
-                types.append(opts_dict["rman_opts"]["types"]["file_sources"]["file_path"][option_name])
-            else:
-                for opts_subdict in opts_dict:
-                    if opts_subdict == "filters":
-                        continue
-                    if option_name in opts_dict[opts_subdict]["types"]:
-                        types.append(opts_dict[opts_subdict]["types"][option_name])
-                # will only enter if we failed to find and return the option value
-                raise logging.warning(f"Given option {option_name} not found.")
-
-        return types
-
-    @classmethod
-    def set_filters(cls, opts_dict: dict, filter_names: Union[str, list], values) -> dict:
-        """Given opts_dict and filter_name, will set filter_name to value
-
-        Args:
-            opts_dict (dict): dictionary of RTCore options. Should follow format of dict returned by cls.get_defaults()
-            filter_names (str or list): name(s) of filter to change
-            values: value to set filter_name to
-        """
-        if type(filter_names) != list:
-            if type(filter_names) == str:
-                filter_names = [filter_names]
-                values = [values]
-            else:
-                raise TypeError(f"Expecting string or list for filter_names. Given {type(filter_names)}")
-        found = False
-        filters_dict = opts_dict["filters"]["values"]
-        for idx, filter_name in enumerate(filter_names):
-            if filter_name in filters_dict:
-                filters_dict[filter_name] = values[idx]
-                found = True
-                continue
-            else:
-                for filter_set in filters_dict:
-                    if type(filters_dict[filter_set]) == dict and filter_name in filters_dict[filter_set]:
-                        filters_dict[filter_set][filter_name] = values[idx]
-                        found = True
-                        # set filter_ligands_flag if we set a ligand filter
-                        if filter_set == "ligand_filters":
-                            filters_dict["filter_ligands_flag"] = True
-            if not found:
-                raise RTCoreError(f"Given filter_name {filter_name} not found. Please check spelling try setting again.")
-
-        return opts_dict
-
-    @classmethod
-    def get_filters(cls, opts_dict: dict, filter_names: Union[str, list]) -> list:
-        """Given opts_dict and filter_name, will set filter_name to value
-
-        Args:
-            opts_dict (dict): dictionary of RTCore options. Should follow format of dict returned by cls.get_defaults()
-            filter_names (str or list): name(s) of filter to change
-            values: value to set filter_name to
-        """
-        if type(filter_names) != list:
-            if type(filter_names) == str:
-                filter_names = [filter_names]
-            else:
-                raise TypeError(f"Expecting string or list for filter_names. Given {type(filter_names)}")
-        filters = []
-        filters_dict = opts_dict["filters"]["values"]
-        print(filters_dict)
-        for filter_name in filter_names:
-            found = False
-            if filter_name in filters_dict:
-                filters.append(filters_dict[filter_name])
-                found = True
-                continue
-            else:
-                for filter_set in filters_dict:
-                    if type(filters_dict[filter_set]) == dict and filter_name in filters_dict[filter_set]:
-                        filters.append(filters_dict[filter_set][filter_name])
-                        found = True
-            if not found:
-                raise RTCoreError(f"Given filter_name {filter_name} not found.")
-
-        return filters
-
-    @classmethod
-    def get_filter_types(cls, opts_dict: dict, filter_names: Union[str, list]) -> list:
-        """Given opts_dict and filter_name, will set filter_name to value
-
-        Args:
-            opts_dict (dict): dictionary of RTCore options. Should follow format of dict returned by cls.get_defaults()
-            filter_names (str or list): name(s) of filter to change
-            values: value to set filter_name to
-        """
-        if type(filter_names) != list:
-            if type(filter_names) == str:
-                filter_names = [filter_names]
-            else:
-                raise TypeError(f"Expecting string or list for filter_names. Given {type(filter_names)}")
-        filters_types = []
-        filters_dict = opts_dict["filters"]["types"]
-        for filter_name in filter_names:
-            found = False
-            if filter_name in filters_dict:
-                filters_types.append(filters_dict[filter_name])
-                found = True
-                continue
-            else:
-                for filter_set in filters_dict:
-                    if type(filters_dict[filter_set]) == dict and filter_name in filters_dict[filter_set]:
-                        filters_types.append(filters_dict[filter_set][filter_name])
-                        found = True
-            if not found:
-                raise RTCoreError(f"Given filter_name {filter_name} not found.")
-
-        return filters_types
 
     def __enter__(self):
         return self
@@ -396,7 +132,6 @@ class RingtailCore:
         ):
             self.storageman.check_storage_empty()
         logging.info("Adding results...")
-        self.results_man = ResultsManager(**self.rman_opts)
         self.results_man.process_results()
 
     def save_receptors(self, receptor_file):
@@ -471,8 +206,6 @@ class RingtailCore:
         import subprocess
         from rdkit.Chem import PyMol
 
-        self._prepare_output_manager()
-
         # launch pymol session
         p = subprocess.Popen(
         ["pymol", "-R"],
@@ -525,10 +258,10 @@ class RingtailCore:
         """
 
         logging.info("Filtering results...")
-        self._prepare_output_manager()
+        self.output_manager.create_log_file()
         # get possible permutations of interaction with max_miss excluded
         interaction_combs = self._generate_interaction_combinations(
-            self.filters["max_miss"]
+            self.filters.max_miss
         )
 
         for ic_idx, combination in enumerate(interaction_combs):
@@ -558,7 +291,7 @@ class RingtailCore:
         """Get data requested in self.out_opts['outfields'] from the
         results view of a previous filtering
         """
-        self._prepare_output_manager()
+        self.output_manager.create_log_file()
         new_data = self.storageman.fetch_data_for_passing_results(
             self.out_opts["outfields"]
         )
@@ -571,7 +304,6 @@ class RingtailCore:
         """
 
         logging.info("Creating plot of results")
-        self._prepare_output_manager()
         # get data from storageMan
         all_data, passing_data = self.storageman.get_plot_data()
         all_plot_data_binned = dict()
@@ -615,13 +347,12 @@ class RingtailCore:
             "le_percentile",
         ]
 
-        property_filters = self.filters["properties"]
         for key in properties_keys:
-            if property_filters[key] is not None:
-                filters_list.append((key, property_filters[key]))
+            if getattr(self.filters, key) is not None:
+                filters_list.append(getattr(self.filters, key))
 
         # interaction filters
-        interaction_filters = self.filters["interactions"]
+        interaction_filters = {"V": self.filters.vdw_interactions, "H": self.filters.hb_interactions, "R": self.filters.reactive_interactions}
         for key in interaction_filters:
             if interaction_filters[key] is not None:
                 kept_interactions = []
@@ -632,12 +363,11 @@ class RingtailCore:
                 filters_list.append((key, kept_interactions))
 
         # get interaction count filters
-        interact_count_filters = self.filters["interactions_count"]
-        for count in interact_count_filters:
+        for count in self.filters.interactions_count:
             filters_list.append(count)  # already a tuple, don't need to reformat
 
         # add react_any flag
-        filters_list.append(("react_any", self.filters["react_any"]))
+        filters_list.append(("react_any", self.filters.react_any))
 
         return filters_list
 
@@ -648,7 +378,6 @@ class RingtailCore:
             write_nonpassing (bool, optional): Option to include non-passing poses for passing ligands
         """
 
-        self._prepare_output_manager()
         if not self.storageman.check_passing_view_exists():
             logging.warning(
                 "Given results bookmark does not exist in database. Cannot write passing molecule SDFs"
@@ -855,12 +584,6 @@ class RingtailCore:
             .replace(" \n", "")
         )
 
-    def _prepare_output_manager(self):
-        if self.output_manager is None:
-            self.output_manager = OutputManager(
-                self.out_opts["log"], self.out_opts["export_sdf_path"]
-            )
-
     def _generate_pdbqt_block(self, pdbqt_lines):
         """Generate pdbqt block from given lines from a pdbqt
 
@@ -955,7 +678,7 @@ class RingtailCore:
         """
 
         all_interactions = []
-        for _type, interactions in self.filters["interactions"].items():
+        for _type, interactions in {"V": self.filters.vdw_interactions, "H": self.filters.hb_interactions, "R": self.filters.reactive_interactions}.items():
             for interact in interactions:
                 all_interactions.append(_type + "-" + interact[0])
 
@@ -978,299 +701,3 @@ class RingtailCore:
             return combinations + self._generate_interaction_combinations(
                 max_miss=max_miss - 1
             )
-
-
-# # # # # # # # # # # # # #
-# # # Output class # # #
-# # # # # # # # # # # # # #
-
-
-class OutputManager:
-    """Class for creating outputs
-
-    Attributes:
-        ad_to_std_atomtypes (dictionary): Mapping of autodock atomtypes (key)
-            to standard PDB atomtypes (value). Initialized as None, will be
-            loaded from AD_to_STD_ATOMTYPES.json if required
-        ax (pyplot axis): Axis for scatterplot of LE vs Energy
-        conformer_indices (list): Description
-        fig_base_name (str): Name for figures excluding file extension
-        flex_residue_smiles (Dictionary): Contains smiles used to create
-            RDKit objects for flexible residues
-        log (string): name for log file
-        vsman (VSManager): VSManager object that created output_manager
-
-    """
-
-    def __init__(self, log_file="output_log.txt", export_sdf_path=""):
-        """Initialize OutputManager object and create log file
-
-        Args:
-            log_file (string): name for log file
-            export_sdf_path (string): path for exporting sdf files
-        """
-
-        self.log = log_file
-        self.export_sdf_path = export_sdf_path
-
-        self._create_log_file()
-
-    def plot_all_data(self, binned_data):
-        """takes dictionary of binned data where key is the
-        coordinates of the bin and value is the number of points in that bin.
-        Adds to scatter plot colored by value
-
-        Args:
-            binned_data (dict): Keys are tuples of key and y value for bin.
-                Value is the count of points falling into that bin.
-        """
-        try:
-            # gather data
-            energies = []
-            leffs = []
-            bin_counts = []
-            for data_bin in binned_data.keys():
-                energies.append(data_bin[0])
-                leffs.append(data_bin[1])
-                bin_counts.append(binned_data[data_bin])
-
-            # start with a square Figure
-            fig = plt.figure()
-
-            gs = fig.add_gridspec(
-                2,
-                2,
-                width_ratios=(7, 2),
-                height_ratios=(2, 7),
-                left=0.1,
-                right=0.9,
-                bottom=0.1,
-                top=0.9,
-                wspace=0.05,
-                hspace=0.05,
-            )
-
-            self.ax = fig.add_subplot(gs[1, 0])
-            ax_histx = fig.add_subplot(gs[0, 0], sharex=self.ax)
-            ax_histy = fig.add_subplot(gs[1, 1], sharey=self.ax)
-            fig.colorbar(
-                mappable=cm.ScalarMappable(
-                    colors.Normalize(vmin=min(bin_counts), vmax=max(bin_counts))
-                ),
-                label="Scatterplot bin count",
-            )
-            self.ax.set_xlabel("Best docking score / kcal/mol")
-            self.ax.set_ylabel("Best Ligand Efficiency")
-        except Exception as e:
-            raise OutputError("Error occurred while initializing plot") from e
-
-        self.scatter_hist(energies, leffs, bin_counts, self.ax, ax_histx, ax_histy)
-
-    def plot_single_point(self, x, y, color="black"):
-        """Add point to scatter plot with given x and y coordinates and color.
-
-        Args:
-            x (float): x coordinate
-            y (float): y coordinate
-            color (str, optional): Color for point. Default black.
-        """
-        try:
-            self.ax.scatter([x], [y], c=color)
-        except Exception as e:
-            raise OutputError("Error occurred while plotting") from e
-
-    def save_scatterplot(self):
-        """
-        Saves current figure as scatter.png
-        """
-        try:
-            plt.savefig("scatter.png", bbox_inches="tight")
-            plt.close()
-        except Exception as e:
-            raise OutputError("Error while saving figure") from e
-
-    def write_log(self, lines):
-        """Writes lines from results iterable into log file
-
-        Args:
-            lines (iterable): Iterable with tuples of data for
-                writing into log
-        """
-        try:
-            for line in lines:
-                logging.info(line)
-                self._write_log_line(
-                    str(line).replace("(", "").replace(")", "")
-                )  # strip parens from line, which is natively a tuple
-            self._write_log_line("***************\n")
-        except Exception as e:
-            raise OutputError("Error occurred during log writing") from e
-
-    def _write_log_line(self, line):
-        """write a single row to the log file
-
-        Args:
-            line (string): Line to write to log
-        """
-        try:
-            with open(self.log, "a") as f:
-                f.write(line)
-                f.write("\n")
-        except Exception as e:
-            raise OutputError(f"Error writing line {line} to log") from e
-
-    def log_num_passing_ligands(self, number_passing_ligands):
-        """
-        Write the number of ligands which pass given filter to log file
-
-        Args:
-            number_passing_ligands (int): number of ligands that passed filter
-        """
-        try:
-            with open(self.log, "a") as f:
-                f.write("\n")
-                f.write(
-                    "Number passing ligands: {num} \n".format(
-                        num=str(number_passing_ligands)
-                    )
-                )
-                f.write("---------------\n")
-        except Exception as e:
-            raise OutputError("Error writing number of passing ligands in log") from e
-
-    def write_results_bookmark_to_log(self, bookmark_name):
-        """Write the name of the result bookmark into log
-
-        Args:
-            bookmark_name (string): name of current results' bookmark in db
-        """
-        try:
-            with open(self.log, "a") as f:
-                f.write("\n")
-                f.write(f"Result bookmark name: {bookmark_name}\n")
-        except Exception as e:
-            raise OutputError("Error writing bookmark name to log") from e
-
-    def write_out_mol(self, ligname, mol, flexres_mols, properties):
-        """writes out given mol as sdf
-
-        Args:
-            ligname (string): name of ligand that will be used to
-                name output SDF file
-            mol (RDKit mol object): RDKit molobject to be written to SDF
-            flexres_mols (list): dictionary of rdkit molecules for flexible residues
-            properties (dict): dictionary of list of properties to add to mol before writing
-        """
-        try:
-            filename = self.export_sdf_path + ligname + ".sdf"
-            mol_flexres_list = [mol]
-            mol_flexres_list += flexres_mols
-            mol = RDKitMolCreate.combine_rdkit_mols(mol_flexres_list)
-            # convert properties to strings as needed
-            for k, v in properties.items():
-                if isinstance(v, list):
-                    v = json.dumps(v)
-                elif not isinstance(v, str):
-                    v = str(v)
-                mol.SetProp(k, v)
-
-            with SDWriter(filename) as w:
-                for conf in mol.GetConformers():
-                    w.write(mol, conf.GetId())
-
-        except Exception as e:
-            raise OutputError("Error occurred while writing SDF from RDKit Mol") from e
-
-    def _create_log_file(self):
-        """
-        Initializes log file
-        """
-        try:
-            with open(self.log, "w") as f:
-                f.write("Filtered poses:\n")
-                f.write("***************\n")
-        except Exception as e:
-            raise OutputError("Error while creating log file") from e
-
-    def scatter_hist(self, x, y, z, ax, ax_histx, ax_histy):
-        """
-        Makes scatterplot with a histogram on each axis
-
-        Args:
-            x (list): x coordinates for data
-            y (list): y coordinates for data
-            z (list): z coordinates for data
-            ax (matplotlib axis): scatterplot axis
-            ax_histx (matplotlib axis): x histogram axis
-            ax_histy (matplotlib axis): y histogram axis
-        """
-        try:
-            # no labels
-            ax_histx.tick_params(axis="x", labelbottom=False)
-            ax_histy.tick_params(axis="y", labelleft=False)
-
-            # the scatter plot:
-            ax.scatter(x, y, c=z, cmap="viridis")
-
-            # now determine nice limits by hand:
-            xbinwidth = 0.25
-            ybinwidth = 0.01
-            xminlim = (int(min(x) / xbinwidth) + 3) * xbinwidth
-            xmaxlim = (int(max(x) / xbinwidth) + 3) * xbinwidth
-            yminlim = (int(min(y) / ybinwidth) + 3) * ybinwidth
-            ymaxlim = (int(max(y) / ybinwidth) + 3) * ybinwidth
-
-            xbins = np.arange(xminlim, xmaxlim + xbinwidth, xbinwidth)
-            ybins = np.arange(yminlim, ymaxlim + ybinwidth, ybinwidth)
-
-            ax_histx.hist(x, bins=xbins)
-            ax_histy.hist(y, bins=ybins, orientation="horizontal")
-        except Exception as e:
-            raise OutputError("Error occurred while adding all data to plot") from e
-
-    def write_filters_to_log(self, filters_dict, included_interactions):
-        """Takes dictionary of filters, formats as string and writes to log file
-
-        Args:
-            filters_dict (dict): dictionary of filtering options
-        """
-        try:
-            buff = ["##### PROPERTIES"]
-            for k, v in filters_dict["properties"].items():
-                if v is not None:
-                    v = "%2.3f" % v
-                else:
-                    v = " [ none ]"
-                buff.append("#  % 7s : %s" % (k, v))
-            if filters_dict["filter_ligands_flag"]:
-                buff.append("#### LIGAND FILTERS")
-                for k, v in filters_dict["ligand_filters"].items():
-                    if v is not None:
-                        if isinstance(v, list):
-                            v = ", ".join([f for f in v if f != ""])
-                    else:
-                        v = " [ none ]"
-                    buff.append("#  % 7s : %s" % (k, v))
-            buff.append("#### INTERACTIONS")
-            labels = ["~", ""]
-            for _type, info in filters_dict["interactions"].items():
-                kept_interactions = []
-                if len(info) == 0:
-                    buff.append("#  % 7s :  [ none ]" % (_type))
-                    continue
-                for interact in info:
-                    if _type + "-" + interact[0] not in included_interactions:
-                        continue
-                    else:
-                        kept_interactions.append(interact)
-                res_str = ", ".join(
-                    ["(%s)%s" % (labels[int(x[1])], x[0]) for x in kept_interactions]
-                )
-                l_str = "#  % 7s : %s" % (_type, res_str)
-                buff.append(l_str)
-
-            for line in buff:
-                self._write_log_line(line)
-
-        except Exception as e:
-            raise OutputError("Error occurred while writing filters to log") from e
