+++ test/test_units.py
@@ -37,72 +37,3 @@ class Test_StorageManSQLite:
         assert bookmark_filters_db_str == json.dumps(filters)
         cur.close()
         conn.close()
-
-    def test_set_opts(self):
-        opts = RingtailCore.get_defaults()
-        RingtailCore.set_opts(opts, "storage_type", "sqlite")
-        RingtailCore.set_opts(opts, "path", [["test_data/"]])
-        RingtailCore.set_opts(opts, "recursive", True)
-
-        assert opts["storage_opts"]["values"]["storage_type"] == "sqlite"
-        assert opts["rman_opts"]["values"]["file_sources"]["file_path"]["path"] == [["test_data/"]]
-        assert opts["rman_opts"]["values"]["file_sources"]["file_path"]["recursive"] == True
-
-        opts2 = RingtailCore.get_defaults()
-        RingtailCore.set_opts(opts2, ["storage_type","path","recursive"], ["sqlite", [["test_data/"]], True])
-
-        assert opts2["storage_opts"]["values"]["storage_type"] == "sqlite"
-        assert opts2["rman_opts"]["values"]["file_sources"]["file_path"]["path"] == [["test_data/"]]
-        assert opts2["rman_opts"]["values"]["file_sources"]["file_path"]["recursive"] == True
-
-    def test_get_opts(self):
-        opts = RingtailCore.get_defaults()
-        opts["storage_opts"]["values"]["storage_type"] = "sqlite"
-        opts["rman_opts"]["values"]["file_sources"]["file_path"]["path"] = [["test_data/"]]
-        opts["rman_opts"]["values"]["file_sources"]["file_path"]["recursive"] = True
-
-        assert RingtailCore.get_opts(opts, "storage_type")[0] == "sqlite"
-        assert RingtailCore.get_opts(opts, "path")[0] == [["test_data/"]]
-        assert RingtailCore.get_opts(opts, "recursive")[0] == True
-
-        opts2 = RingtailCore.get_defaults()
-        opts2["storage_opts"]["values"]["storage_type"] = "sqlite"
-        opts2["rman_opts"]["values"]["file_sources"]["file_path"]["path"] = [["test_data/"]]
-        opts2["rman_opts"]["values"]["file_sources"]["file_path"]["recursive"] = True
-
-        assert RingtailCore.get_opts(opts2, ["storage_type", "path", "recursive"]) == ["sqlite", [["test_data/"]], True]
-
-    def test_set_filters(self):
-        opts = RingtailCore.get_defaults()
-        RingtailCore.set_filters(opts, "eworst", -3.0)
-        assert opts["filters"]["values"]["properties"]["eworst"] == -3.0
-        RingtailCore.set_filters(opts, "V", ["cucumber"])
-        assert opts["filters"]["values"]["interactions"]["V"] == ["cucumber"]
-        RingtailCore.set_filters(opts, "N", ["ligand1"])
-        assert opts["filters"]["values"]["ligand_filters"]["N"] == ["ligand1"]
-        assert opts["filters"]["values"]["filter_ligands_flag"] == True
-        RingtailCore.set_filters(opts, "max_miss", 3)
-        assert opts["filters"]["values"]["max_miss"] == 3
-
-        opts2 = RingtailCore.get_defaults()
-        RingtailCore.set_filters(opts2, ["eworst", "V", "N", "max_miss"], [-3.0, ["cucumber"], ["ligand1"], 3])
-        assert opts2["filters"]["values"]["properties"]["eworst"] == -3.0
-        assert opts2["filters"]["values"]["interactions"]["V"] == ["cucumber"]
-        assert opts2["filters"]["values"]["ligand_filters"]["N"] == ["ligand1"]
-        assert opts2["filters"]["values"]["filter_ligands_flag"] == True
-        assert opts2["filters"]["values"]["max_miss"] == 3
-
-    def test_get_filters(self):
-        opts = RingtailCore.get_defaults()
-        
-        opts["filters"]["values"]["properties"]["eworst"] = -3.0
-        opts["filters"]["values"]["interactions"]["V"] = ["cucumber"]
-        opts["filters"]["values"]["ligand_filters"]["N"] = ["ligand1"]
-        opts["filters"]["values"]["max_miss"] = 3
-
-        assert RingtailCore.get_filters(opts, "eworst")[0] == -3.0
-        assert RingtailCore.get_filters(opts, "V")[0] == ["cucumber"]
-        assert RingtailCore.get_filters(opts, "N")[0] == ["ligand1"]
-        assert RingtailCore.get_filters(opts, "max_miss")[0] == 3
-
-        assert RingtailCore.get_filters(opts, ["eworst", "V", "N", "max_miss"]) == [-3.0, ["cucumber"], ["ligand1"], 3]
